Extracted ImageUploadButton as separate plugin.
